chore: rename no-components files with camelCase
