+++ v-marionette-renewal/src/components/App.tsx
@@ -4,7 +4,7 @@ import theme from 'styles/theme'
 import LeftWindow from 'components/leftwindow/LeftWindow'
 import { uiStores } from 'stores/uiStores'
 import { VRMSceneScreen } from 'components/VRMSceneScreen'
-import { startMpActions, stopMpActions } from 'models/MediapipeAction'
+import { startMpActions, stopMpActions } from 'models/mediapipeAction'
 import Footer from 'components/Footer'
 import { autorun } from 'mobx'
 import { avatar } from 'models/avatar'
