+++ v-marionette-renewal/src/components/VRMSceneScreen.tsx
@@ -2,7 +2,7 @@ import * as THREE from 'three'
 import React from 'react'
 import { IReactionDisposer, reaction } from 'mobx'
 import styled from '@emotion/styled'
-import networkHandler from 'models/NetworkHandler'
+import networkHandler from 'models/networkHandler'
 import { uiStores } from 'stores/uiStores'
 import { Viewer, Avatar, avatar } from 'models/avatar/'
 import { throttle } from 'lodash'
