+++ v-marionette-renewal/src/components/leftwindow/NetworkSettingWindow.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Stack, Button } from '@mui/material'
 import { networkSettings } from 'stores/userSettings'
 import VMTextField from 'components/leftwindow/VMTextField'
-import networkHandler from 'models/NetworkHandler'
+import networkHandler from 'models/networkHandler'
 import { IReactionDisposer, reaction } from 'mobx'
 
 const NetworkSettingWindow: React.FC = () => {
