+++ v-marionette-renewal/src/models/avatar/IK/ikTargetTracker.ts
@@ -1,7 +1,7 @@
 import { GLTFNode, VRM, VRMSchema } from '@pixiv/three-vrm'
 import * as THREE from 'three'
 import { aiRim } from 'types'
-import * as IKSolver from './IKSolver'
+import * as IKSolver from './iKSolver'
 import { localPosToWorldPos } from 'models/utils'
 
 export class IkTargetTracker {
