+++ v-marionette-renewal/src/models/avatar/IK/index.ts
@@ -1,1 +1,1 @@
-export * from './VrmIK'
+export * from './vrmIk'
