+++ v-marionette-renewal/src/models/avatar/rotateHand.ts
