+++ v-marionette-renewal/src/models/networkHandler.ts
